bike-lending-app/sql/create-bikes.sql: Change "bikesDB" to "bikesdb".
When running
    mvn datanucleus:schema-create
An error is returned
    An error was encountered creating a PersistenceManagerFactory : Access denied for user
    'spq'@'localhost' to database 'bikesdb' - please consult the log for more information.

As can be seen in the error message above there is attempted access to the
"bikesdb" database which does not exist as the name specified in create-bikes.sql
is "bikesDB". This is not an issue on Windows systems where paths
are case-insensitive. On Unix-like systems the paths are case-sensitive
leading to the error above.

To fix this issue all mentions of "bikesDB" are replaced with "bikedb".